generalise and remove unneeded constants
