+++ packages/client/src/BaseConversation.ts
@@ -1,4 +1,10 @@
-import { Callbacks, Mode, Status } from "@elevenlabs/types";
+import {
+  Callbacks,
+  ELEVENLABS_CONVERSATION_SYMBOL,
+  type ElevenLabsConversationAPI,
+  Mode,
+  Status,
+} from "@elevenlabs/types";
 import type {
   BaseConnection,
   DisconnectionDetails,
@@ -129,6 +135,51 @@ export class BaseConversation {
     this.connection.onDisconnect(this.endSessionWithDetails);
     this.connection.onModeChange(mode => this.updateMode(mode));
     this.updateStatus("connected");
+
+    if (this.options.debug) {
+      this.exposeGlobalAPI();
+    }
+  }
+
+  private exposeGlobalAPI() {
+    if (typeof window === "undefined") return;
+
+    const self = this;
+    const api: ElevenLabsConversationAPI = {
+      get status() {
+        return self.status;
+      },
+      get conversationId() {
+        return self.connection.conversationId;
+      },
+      get inputFormat() {
+        return self.connection.inputFormat;
+      },
+      sendUserMessage(text: string) {
+        self.sendUserMessage(text);
+      },
+      sendAudio(base64Audio: string, sampleRate?: number) {
+        return self.sendAudioToConversation(base64Audio, sampleRate);
+      },
+    };
+
+    (window as any)[ELEVENLABS_CONVERSATION_SYMBOL] = api;
+  }
+
+  private cleanupGlobalAPI() {
+    if (typeof window !== "undefined") {
+      delete (window as any)[ELEVENLABS_CONVERSATION_SYMBOL];
+    }
+  }
+
+  /** Override in subclasses that support audio injection. */
+  protected sendAudioToConversation(
+    _base64Audio: string,
+    _sampleRate?: number
+  ): Promise<{ cancel: () => void }> {
+    return Promise.reject(
+      new Error("sendAudio is not available for this conversation type")
+    );
   }
 
   public endSession() {
@@ -146,6 +197,9 @@ export class BaseConversation {
   };
 
   protected async handleEndSession() {
+    if (this.options.debug) {
+      this.cleanupGlobalAPI();
+    }
     this.connection.close();
   }
